WIP: moved example_driver into testing_util namespace
