+++ google/cloud/storage/examples/storage_examples_common.h
@@ -24,11 +24,11 @@ namespace cloud {
 namespace storage {
 namespace examples {
 
-using ::google::cloud::internal::CheckEnvironmentVariablesAreSet;
-using ::google::cloud::internal::Commands;
-using ::google::cloud::internal::CommandType;
-using ::google::cloud::internal::Example;
-using ::google::cloud::internal::Usage;
+using ::google::cloud::testing_util::CheckEnvironmentVariablesAreSet;
+using ::google::cloud::testing_util::Commands;
+using ::google::cloud::testing_util::CommandType;
+using ::google::cloud::testing_util::Example;
+using ::google::cloud::testing_util::Usage;
 
 bool UsingTestbench();
